refactor(jmespath_functions): Move into shared
